code optimization

+++ JavaBeginners01/src/defaultt/CurrencyConverter.java
@@ -3,24 +3,24 @@ package defaultt;
 public class CurrencyConverter {
 
 	public static void main(String[] args) {
-		double dollarrate = 1500.56;
-		double poundrate = 1729.44;
-		double eurorate = 2188.84;
-		double dollaramount = 457.62;
-		double poundamount = 683.28;
-		double euroamount = 888.33;
-		double nairaamount = 3860547.9876;
-		double dollartonaira = dollaramount * dollarrate;
+		float dollarrate = 1500.56f;
+		float poundrate = 1729.44f;
+		float eurorate = 2188.84f;
+		float dollaramount = 457.62f;
+		float poundamount = 683.28f;
+		float euroamount = 888.33f;
+		float nairaamount = 3860547.9876f;
+		float dollartonaira = dollaramount * dollarrate;
 		System.out.println(dollaramount +" dollars to naira is " + dollartonaira);
-		double poundtonaira = poundamount * poundrate;
+		float poundtonaira = poundamount * poundrate;
 		System.out.println(poundamount + " pounds to naira is " + poundtonaira);
-		double eurotonaira = euroamount * eurorate;
+		float eurotonaira = euroamount * eurorate;
 		System.out.println(euroamount + " euros to naira is " + eurotonaira);
-		double nairatodollars = nairaamount/dollarrate;
+		float nairatodollars = nairaamount/dollarrate;
 		System.out.println(nairaamount + " naira to dollars is " + nairatodollars);
-		double nairatopounds = nairaamount/poundrate;
+		float nairatopounds = nairaamount/poundrate;
 		System.out.println(nairaamount + " naira to pounds is " + nairatopounds);
-		double nairatoeuros = nairaamount/eurorate;
+		float nairatoeuros = nairaamount/eurorate;
 		System.out.println(nairaamount + " naira to euros is " + nairatoeuros);
